Remove last updated bits from docs
This adds nothing and is wrong. The author is not the person who made
changes to that file last. The time is just most recent commit time, not
last modification of that file.

This also makes the build output uncacheable and is the primary reason
that the commit history of gh-pages is as large as it is. Virtually
every single file changes on every build because of this little bit of
content.