Move resolver specific code out of schema-extractor
Summary: This is a quick re-organization of some of the resolver code to move the data structures and function specific to resolvers out of the schema-extractor crate. The functions left in that crate are all related directly to the hermes_estree AST. This could arguably be moved into the resolver crate but could be useful if we need to share these functions with other JS parsing in the future.

Reviewed By: tyao1

Differential Revision: D57231255

fbshipit-source-id: 69bc715ece8fc4b841100fcd4e6607a44495a52b