Factor out entity name extractor code to a function
Summary:
This is another organizational diff to move out the code that matches on a fields entity name to it's own function. As much as possible, I want to put the code that does matching on the `FlowAnnotationType` into the `extract_*` functions to preserve the high level intent of the other schema generation code.

I also fixed the fallthrough error message to be more general.

Reviewed By: tyao1

Differential Revision: D57238848

fbshipit-source-id: 50addfb9601e97fa9ae984b8cb9a891b6a480212