Implement schema generation for weak object fields
Summary: This diff implements the ability to add fields to weak objects in the same mannger as adding a field to a strong object. We realized that, while weak object literal types were parsed into the IR, the runtime is not currently generated. We also want the ability for custom function implementation of weak object fields. This diff also disables the previous implementation of parsing object literals until the runtime JS can also be generated.

Reviewed By: tyao1

Differential Revision: D58501581

fbshipit-source-id: d8df5f0f270b5669fe6cacf129671ffd29804f06