Add correct root fragment information
Summary: I noticed when comparing the IR output in our tests, that the root fragment information for the Flow based schema generation was incorrect compared to the docblock based schema gen. This diff updates them so they are using the same information. This is important for the location of error messages.

Reviewed By: captbaritone

Differential Revision: D58709712

fbshipit-source-id: 457a1ced4f101660b2a0fa7cef3aa3c56fac4d48