Have _implementsInterface itself report if its missing data
Summary:
In some cases, like a graph relationship change, we might try to check a type condition on a concrete type where we've never fetch the `__is[AbstractType]` flag for that type. In this case we need to report that we are missing data, in case that field is still in flight.

This consolidates that check into our generic `_implementsInterface` method.

Reviewed By: evanyeung

Differential Revision: D58742672

fbshipit-source-id: 8698084de4c6acc46b0511897991874a81de9b1e