Fix behavior of @alias on inline fragments on abstract types that don't match
Summary:
If you look comment at the top of the _readInlineFragment method, you can see that it says that we should return null if we _know_ we have not matched and undefined if we don't know (due to missing data).

We implemented this correctly for concrete types, but we had it backwards for abstract types. Return values other than `false` are only used by `alias` so this should be a safe change since there are ~0 uses of `alias` so far in the wild, and the change of undefined vs null is very minimal.

Reviewed By: DataDataDataViz

Differential Revision: D58742671

fbshipit-source-id: 00620efedd6fae9f46afabe4c32e3d4bf7f83d47