Avoid reading inline fragments if abstract type condition does not match
Summary:
For legacy reasons (which I'll revisit in a separate conversation) we _always_ try to read inline fragments on abstract types, even if we know the type condition did not match.

For aliased fragments this behavior is obviously broken, and we also have an opportunity chance to fix the behavior since we don't have any legacy use cases. This diff has aliased inline fragments on abstract types respect the type condition

Reviewed By: evanyeung

Differential Revision: D58742669

fbshipit-source-id: 491efdb1051e42395ff9a4d8c2769afdd75e1d37