Map Flow -> GraphQL Scalars in schema gen
Summary:
This diff adds support for mapping Flow types defined in the arguments or return type of client side resolvers back to GraphQL custom scalars based on the config file. It takes the custom scalar map from the config file which maps from custom scalars to Flow types and inverts it. Then, if it sees that a resolver field uses a Flow type in that custom scalar map, it replaces the Flow type with the corresponding custom scalar.

The actual logic in this diff is fairly simple in that it just adds a lookup map to the only place that we parse Flow types: `return_type_to_type_annotation`. However, it required a bunch of shuffling of code to 1. thread the custom scalars from the config through to the schema gen code and 2. instead of putting the Flow type in the unresolved field struct, we now put the GraphQL type. The types for the unresolved fields have to be worked out earlier because the custom scalar code needs access to the imports of the file the resolver is defined in.

Reviewed By: tyao1

Differential Revision: D58981404

fbshipit-source-id: 048bc5b9c380f2efd5b121837534b3f20948f31c