Merged PR 1030517: Add option to only generate for the current directory to prevent writing outside
In 1JS we have the issue that writing cross packages is discouraged/not allowed.

However we do have the need to run the relay compiler across packages inside the repo.

The compiler will need to know about all the files across the packages however we want the possibility to only generate files inside the folder where the compiler is actually run from.

This PR adds a compiler flag `--write-current-dir-only` that will make skip writing files if they are not inside the current folder.

```
/midgard
   /packages
       /dependency
           /src/ReactComponentWithFragment.tsx
           package.json
         /my-app
           /src/AppComponentWithQuery.tsx
           package.json
```

When running the compiler in `/midgard/packages/my-app` it should read files from `/midgard/packages/dependency` but not write files in that location.

We'll generate the files for those packages separately when needed.

Initially we worked around this by using `artifactDirectory` in the `my-app` package so that all generated files end up there, but this isn't great as we loose colocation of the generated files where this is used.