Refactor alias to always use inline fragments (and compose with other features)
Summary:
As I worked to enable `alias` compatibility with other relay features that are implemented as directives on fragment spreads, I realized the biggest challenge was ensuring nesting order was correct. For ~all other features we want the alias to wrap the actual feature.

This made me realize that we can just always transform `alias` into an inline fragment that wraps the un-aliased fragment spread. This has a number of advantages:

1. Allows us to simplify the runtime by only handling the inline fragment case (see next diff)
2. Enables most other features to work without needing any change to consider how they interact with alias, and also allows codegen to not have to consider how various features interact
3. Allows us to simplify codegen/typegen to only think about the inline fragment case

But it also has some downsides:

1. We need the alias transfrom to apply before some of the transforms that are applied in common transforms so that nesting happens correctly. However, we also need to ensure the alias transform does not impact normalizaiton or query artifacts. To work around this we need a second transform that strips the wrapper inline fragments that we add around fragment spreads.

Reviewed By: lynnshaoyu

Differential Revision: D58790868

fbshipit-source-id: 8d3f398dc27f2039f1017c4c2191858a85491069