Merged PR 1041636: Skip check for source control if --noSourceControl is passed
This PR adds the option to pass `--noSourceControl` which will skip the `git status` and `sl --version` commands when creating the configuration.

Those checks are run to help the compiler add/remove files from source control for you.
However in cases where the generated files are not checked into source control this isn't helpful, especially when `git status` is "slow", skipping the checks can give a nice boost in compiler performance.

I noticed a significant decrease in performance between 16 and 17 when `git status` was introduced.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature that allows skipping the check for source control when the `--noSourceControl` flag is passed.

#### PR Summary
The changes in this pull request primarily involve the addition of a new command-line flag `--noSourceControl` and the conditional execution of source control checks based on this flag. The version of the `relay-compiler` package has also been updated.
- In `relay-bin/src/main.rs`, a new command-line flag `no_source_control` is introduced and the source control check is made conditional based on this flag.
- In `relay-compiler/src/config.rs`, the source control check during the configuration setup is removed.
- The version of the `relay-compiler` package in `relay-compiler/package.json` is updated from `17.0.0-microsoft.1` to `17.0.0-microsoft.2`.