Skip check for source control integration if --noSourceControl is passed
This PR adds the option to pass `--noSourceControl` which will skip the `git status` and `sl --version` commands when creating the configuration.

Those checks are run to help the compiler add/remove files from source control for you.
However in cases where the generated files are not checked into source control this isn't helpful, especially when `git status` is "slow".

I noticed a significant decrease in performance between 16 and 17 when `git status` was introduced.