Merged PR 1043827: Job name
Job name

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a modification to the job naming convention in the Azure DevOps pipeline.

#### PR Summary
The change in this pull request is focused on improving the job naming in the Azure DevOps pipeline configuration file. This change will make the job names more descriptive and easier to identify.
- In `/.azure-devops/steps/build-compiler-job.yml`, the job name has been updated to include the target parameter, making it more descriptive and easier to identify.