Prevent using @alias within @match
Summary:
`match` expects all of its selections to be `module` fragment spreads. In order for these to be passed to `<MatchContainer>` in the way users expect, we need to ensure that none of those `module` fragments have been pushed into their own namespace.

This validation prevents that.

I also took this oppertunity to improve a few other things I noticed:

1. Don't panic if user passes a non-literal key argument to `match` (user input should never be able to panic the compiler)
2. Improve location reporting for conditional and inline fragments

Reviewed By: tyao1

Differential Revision: D58792831

fbshipit-source-id: be3d0011c24a63f8435773491d20caa4d0395f84