Add support for global custom scalars
Summary:
This diff adds support for using custom scalars for globally defined Flow types. If a globally defined Flow type is found that corresponds to a custom scalar, it does not need to be imported for Relay Resolvers to recognize it and the custom scalar type is substituted.

Also note the shadowing behavior: if a type is imported with the same name as a custom scalar global, the local type is used for instead (in the test case this is a weak object type).

Reviewed By: tyao1

Differential Revision: D59477655

fbshipit-source-id: e3ad14885169ade19add282457103546216c2925