Add method to skip failing tests in OSS, with a warning (#4743)
Summary:
Test failures in OSS are blocking publishing the relay-compiler package to npm and OSS contributions. We can skip the failing GitHub CI tests and add a TODO as we are fixing them long-term in OSS. (cc captbaritone)

Pull Request resolved: https://github.com/facebook/relay/pull/4743

Test Plan:
The test is not skipped internally, a warning is emitted about the errors in OSS, and the jest test passes overall.

Test run for useQueryLoader-test.js:

P1481257935

`flow` passes
Test passes in OSS as well

Reviewed By: evanyeung

Differential Revision: D59764108

Pulled By: monicatang

fbshipit-source-id: 28defe54c7d9ae67f661f9d65e491cfe442638da