Organize sidebar for oss site
Summary:
This diff builds on the last one to organize the sidebar of the oss site. I opted to use a very similar structure to the internal one as the catagories still make sense. I collapsed some of the categories that only have one section in them though. These can be re-expanded when more pages are added.

This also only affects `next` right now. I think we need to run the versioning script to add it to v17 but that would rebuild everything into v17 so we can wait until the next version bump.

Reviewed By: DataDataDataViz

Differential Revision: D60271544

fbshipit-source-id: 79763454c483e30fcc351b710c6688e69c2c6b7e