Don't skip ActorChangeWithStream in OSS, it already passes (#4763)
Summary:
This test does not play well with skipIf because it generates a snapshot and Jest sees the skipped snapshot test as out of date and thus failing. But when I tried to run it locally without skipping it, it seems to pass on my machine.
<img width="471" alt="Screenshot 2024-08-05 at 1 17 10 PM" src="https://github.com/user-attachments/assets/2275916e-21a1-4f42-9b34-e09199dbd3ae">

Pull Request resolved: https://github.com/facebook/relay/pull/4763

Test Plan:
Passes in open source:

 {F1792016271}

 {F1792017116}

Reviewed By: tyao1

Differential Revision: D60784559

Pulled By: captbaritone

fbshipit-source-id: 95dadab10f9a7dfd7259fae3a6ea4cd1f679d5da