Support <Activity> in useFragmentInternal
Summary:
Adds an alternate implementation of `useFragmentInternal()` to support React's unstable `<Activity>` API. The main change with Activity is that effects can get detached/re-attached independently of mount/unmount. We found a bug where, after detaching and reattaching effects, the `handleMissedUpdates()` logic in the effect would compare against a stale version of the previous state, and therefore always think that data had changed.

The fix here is to switch the existing "handle missed updates and subscribe" effect to be a CRUD-style effect, emulated via a ref plus two effects:
* The ref tracks what state we have already subscribed to
* The "update" effect reacts to state changes, and reconciles the current state against the previously subscribed state. Notably, we now always run this effect whenever the fragment result changes, but will just quickly bail out since we know we're already subscribed to that state (same selector and environment)
* The create/delete effect has no deps, and is used to ensure we always have a subscription on mount/re-attach, and teardown the subscription on unmount/re-attach.

To allow testing, `useFragmentInternal()` checks a new feature flag and delegates to the original implementation (now renamed _CURRENT) or the new one (_EXPERIMENTAL).

Reviewed By: tyao1

Differential Revision: D61996595

fbshipit-source-id: ceeb9a63f5bbbdefaac9d546491c01d510ed460e