Attempt to avoid cycled when expanding updated ids to include impacted resolvers (#4792)
Summary:
We have seen a bug internally which appears to be a case where we encounter a cycle in this code and thus an infinite loop. I have yet to figure out how to reproduce that behavior, but it's clear that there is a bug in this code. We maintain a set of visited IDs but then never check that set as you progress.

For now I'm adding what I believe to be a more correct implementation, however I haven't yet been able to reproduce so I'm not 100% sure this will actually fix the issue. Additionally, it's in a sensitive part of the code, so I'm gating it to allow for a gradual rollout and the ability to quickly disable.

Pull Request resolved: https://github.com/facebook/relay/pull/4792

Test Plan: I've ensured the relevant tests validate both paths. During rollout I'll also create Diffs to run e2e tests with the fix enabled.

Differential Revision: D62508943

Pulled By: captbaritone

fbshipit-source-id: c6100c6f4c118f01bf36f2ce04e47825e0f77e93