useQueryLoader() (and refetch()) compatibility with <Activity>
Summary:
Fixes `useQueryLoader()` — and by extension, `usePaginationFragment()` refetch function — to be compatible with React's unstable `<Activity>` API.

The problem is that currently, useQueryLoader frees queries in unmount and reissues the query when remounting. We really don't want to cancel a load if we're just going to use it very soon when the component gets shown again. The fix here is to only free queries in two situations:

* When a newly loaded query commits, we free all other previously loaded queries
* When the component finally unmounts (detected via useInsertionEffect), we free *all* queries.

The final case hits a React bug where useInsertionEffect cleanup isn't running for suspended trees, so we work around this by also scheduling a 5min timeout in normal effect cleanup.

Reviewed By: gordyf, tyao1

Differential Revision: D62592061

fbshipit-source-id: 0c5b05fb51c282a37031fc8e0468758a75d3afb8