Remove now-unnecessary fix that entangles lanes
Summary:
This code was originally added as part of a fix where an update was scheduled on a fragment, but a high-priority render failed to observe that update. The original issue appears to have been fixed, since the test case now passes without the code added to fix it.

In fact, we're seeing the fix cause the opposite problem, where an update gets scheduled at low priority, and a high-priority update flickers ahead to the new state because of this fix, shows the old state, then eventually coalesces on the new state again.

This code is explicitly causing data from different priority updates/lanes to get mixed so let's remove it.

Reviewed By: tyao1

Differential Revision: D62618760

fbshipit-source-id: 3b5f7573179df092736b17157ff3a6bb4f2a80e5