Update Relay API for disk caching
Summary:
This diff adds the basic implementation of disk caching to Relay. It caches on a per-query basis and using a key built with a fixed string, the actor ID, and the operation name. It's designed to work in AsyncStorage on native and localStorage on web. The main idea is to cache the normalized records to disk so that they can be filtered and manipulated before caching. This differs from the network cache for which there is no way to change the records that are stored for the query.

Note: title is a bit unclear internally so that the diff makes sense in OSS.

Adds `getOperationLoader` to the store API so that Relay can cache/load data from disk

Reviewed By: tyao1

Differential Revision: D62275908

fbshipit-source-id: 74b39976f2c6e01e8169cddd3eba2390de9165b0