Use field error events on snapshot (#4798)
Summary:
Previously we had two different object shapes which were nearly the same. One for stashing on the snapshot, and another for logging. By unifying the two shapes we simplify things a bit and avoid the transformation.

As part of this simplification, I've also removed the special case handling for snapshots which contain only missing data errors. This means we will end up with duplicate events if the snapshot is missing multiple fields. This is noisier, but perhaps more correct because it captures the fact that multiple fields were missing. Once/when/if we add runtime pathName to RelayReader, these multiple events will be even more helpful.

Pull Request resolved: https://github.com/facebook/relay/pull/4798

Reviewed By: gordyf

Differential Revision: D63032950

Pulled By: captbaritone

fbshipit-source-id: 30fc941134e641eb31f1440f2d9e066f23f07621