Add feature flag for reader AST module imports
Summary:
This diff adds a feature flag to turn on generating the `moduleImports` field in the Reader AST. This is necessary as, if records are serialized for caching, the module load function cannot be saved. By keeping it in the Reader AST, any 3D module fields can be looked up at read time rather than from the record. It builds upon the recently completed client 3D work.

This diff adds a feature flag to turn on generating the `moduleImports` field in the Reader AST. It builds upon the recently completed client 3D work.

Reviewed By: tyao1

Differential Revision: D62347343

fbshipit-source-id: 2e9f0635f27c112871b45a2fee5c87ff9235eb2c