Avoid running resolver code if root fragment throws with @required(action: THROW) (#4799)
Summary:
I noticed a bug here. In order for the types generated by adding `required(action: THROW)` to a resolver fragment to actually be safe, we need to short circuit and avoid calling the resolver function if the field has thrown.

This lays the ground work for the other bit we need, which is handling throwOnFieldError errors and ensuring we never run a resolver with a fragment that should throw on read.

Pull Request resolved: https://github.com/facebook/relay/pull/4799

Reviewed By: gordyf

Differential Revision: D63032852

Pulled By: captbaritone

fbshipit-source-id: e3b67e15c5b899cd2946856c2689a65ea2ffcf87