Add satisfies type assertion for typescript (#4797)
Summary:
Fixes https://github.com/facebook/relay/issues/4772

Adds a satisfies assertion to the live resolver import:

```
import {User as userRelayModelInstanceResolverType} from "UserTypeResolvers";
// Type assertion validating that `userRelayModelInstanceResolverType` resolver is correctly implemented.
// A type error here indicates that the type signature of the resolver module is incorrect.
(userRelayModelInstanceResolverType satisfies (
  id: User__id$data['id'],
  args: void,
) => LiveState<mixed>);
```

It will also require a minimum Typescript version of TS 4.9 or higher for consumers, which can be checked as we implement https://github.com/facebook/relay/issues/4755. We may have to have add a feature flag to disable this for consumers of the library which do not meet this minimum standard.

Pull Request resolved: https://github.com/facebook/relay/pull/4797

Reviewed By: tyao1

Differential Revision: D63032995

Pulled By: captbaritone

fbshipit-source-id: be909c70f38312cd5931d7824b2e887b7834dc1c