Upstream observeFragment and waitForFragmentData to relay-runtime
Summary:
There are legitimate cases where users may want to read fragment data lazily outside the life-cycle of a React component. Historicaly that was somewhat possible via Environement.lookup + Environment.subscribe (the same primitives used by our React components).

However, over the last several years we've added lots of compelxity to Snapshot, to the point where its not really safe to use those APIs publically anymore, since it requires correctly handling:

- Field errors (required, throwOnFieldError, and logging)
- Resolver field suspense
- Client edge requests
- Plural fragments

Because of this I think it's time for Relay to expose a higher-level API that lets peope lazily (and over time) observe the value of a fragment in the store with all of these details handled.

## Marquee usecase

One critical use case here is the ability to `defer` a fragment that is expensive to fetch, is not needed for rendering, but is accessed inside and event handler. Previously there was not reasonable way to achieve this without forcing some portion of the UI to suspend until that data was avaliable.

## Next steps

We can allow early adopters to try this API and unblock themselves, but to complete it there are still a few thing we must solve:

1. Support plural fragments
2. Support client edge fetches
3. Have a design discussion to deeply consider the implicaitons of adding this API

Reviewed By: tyao1

Differential Revision: D63266248

fbshipit-source-id: 79a83a522bf1daeb2554079a2a1ff02297839d17