Add copy operation command (#4778)
Summary:
I often find myself in a situation where I want to quickly run a Relay query outside of the configured Relay environment, i.e. a GraphQL IDE or another tool. When not using persisted operations, I need to go to the artifact, copy out the value of the `text` property and get rid of all the `\n` line breaks in the text. For persisted operations it's less tedious, I just grab the `id` from the artifact and look up the operation in the operation store, but it's still more cumbersome than it has to be.

This implements a new `relay/printOperation` request in the LSP that returns the transformed operation text of the operation under the cursor or the first operation inside the document. The VS Code Extension is also updated with a `relay.copyOperation` command that invokes that LSP request and copies the operation text to the clipboard.

<img src="https://github.com/user-attachments/assets/52cdff85-dad2-4236-b4d8-c155c02c2c97" width="500" alt="Example of the command in action" />

Pull Request resolved: https://github.com/facebook/relay/pull/4778

Reviewed By: tyao1

Differential Revision: D63033129

Pulled By: captbaritone

fbshipit-source-id: d5caf83b4b0135e4b30ce8cfaf4ef23108f52db6