Stabilize LiveResolverStore as the new implementation of RelayModernStore (#4806)
Summary:
For the last several years we've maintained RelayModernStore and LiveResolverStore in parallel. This was because the the Relay store needed to gain some additional features in order to support [Resolvers](https://relay.dev/docs/next/guides/relay-resolvers/introduction/) and [Live Resolvers](https://relay.dev/docs/next/guides/relay-resolvers/live-fields/).

The upshot was the using Relay Resolvers required adopting a special store. We've now been using this store on ~all surfaces internally for quite a while and have not seen any issues. I believe it should now be safe to unify the implementations and bring the features of the LiveResolver store to the default store: RelayModernStore.

This should allow us to make Relay Resolvers and Live Resolvers as stable (not requiring any additional configuration or feature flags) in an upcoming release. I'll followup with another PR to remove the last feature flag.

Pull Request resolved: https://github.com/facebook/relay/pull/4806

Reviewed By: tyao1

Differential Revision: D63488011

Pulled By: captbaritone

fbshipit-source-id: b70d3fdfb0e37a2c07f890b772588121632741ea