Install @testing-library/react
Summary:
# This diff: Bring in RTL

Adds `testing-library/react` to Relay. This unblocks migrating off of RTR and writing tests that adopt new features such as Actions and Activity.

# Stack

## Problem

We are introducing `testing-library/react` to Relay's test suite to modernize the tests and unblock testing of new features like Actions and Activity. The promise polyfills from React Native (`promise/setImmediate`) and Relay (`promise-polyfill`) block this. There are a couple of issues:

- `setImmediate` is not provided in the browser/`jsdom` environment
- The scheduling of promise callbacks differs between the polyfills and native promises. The polyfills do not play nicely when testing Actions.

Here is a simple Actions test case that should pass but breaks in Relay today:

```js
it.only('simple', async () => {
    let resolve;
    const promise = new Promise(r => {
      resolve = r;
    });

    function TestComponent() {
      const [isPending, startTransition] = useTransition();

      const clickAction = () => {
        startTransition(async () => {
          await promise;
        });
      };

      return (
        <div>
          <button onClick={clickAction}></button>
          <div>{isPending ? 'pending' : 'ready'}</div>
        </div>
      );
    }

    const result = ReactTestLibrary.render(<TestComponent />);

    expect(result.container.textContent).toEqual('ready');

    await act(() => {
      result.getByRole('button').click();
    });

    expect(result.container.textContent).toEqual('pending');

    await act(() => {
      resolve()
    });

    expect(result.container.textContent).toEqual('ready');
  });
```

## Solution

I've tried a bunch of approaches here and this diff stack requires the least amount of overhead and scope creep while unblocking the usage of `jsdom` and native promises in Relay tests.

Originally after speaking with yungsters, we thought it would be great to completely remove the Promise polyfill from fbsource. Promises are built in to Node and Hermes so we shouldn't need it anymore. The problem with this in practice though is that there are scheduling differences between native a polyfilled promise implementations. There are also differences in how `React.act` behaves in each implementation.

These scheduling differences end up breaking lots of tests because we rely on a lot of explicit timing controls such as `act`, `await act`, `jest.runAllImmediates()`, `jest.runAllTimers`, and nested combinations of all of these. When removing the polyfill, we have to restructure how we step through async code in tests. It's not impossible, but its a lot of work to do this across the codebase. It turns this into a BE project much broader than simply unblocking new testing capabilities for Relay.

**The goals here are:**
1. Unblock testing new capabilities in Relay
2. Make test runs between fbsouce and and OSS more consistent
3. Unblock incremental cleanup of polyfilled promises in Relay and in fbsource overall

To do this, we need to stop globally applying RN's polyfill in a few stages. First we split it into its own setup file so it can be optionally included in jest configs. Then we create a new jest config for Relay. A bunch of tests do rely on specific async scheduling, so we can inject the polyfill in just those test modules while stopping the bleeding of it being a global default.

Over time through BE efforts we can clean up the polyfill injections. These tests should be rewritten to migrate away from ReactTestRenderer anyway.

Adds `testing-library/react` to Relay. This unblocks migrating off of RTR and writing tests that adopt new features such as Actions and Activity.

Reviewed By: yungsters

Differential Revision: D63323950

fbshipit-source-id: ecb0cd31af738d61bc27050ed8fea573e99d013d